+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.4.0
+
+* Added addOffset command to LabelEpl.
+* Added get Device to LP2844.
+* Added some type annotations in JSDoc.
+* Tweak label length calculation to be more reliable.
+
 ## 0.3.0
 
 * Added cut command to LP2844 printer class.